Polygon bugs
